+++ components/SdImageComp.tsx
@@ -8,6 +8,7 @@ import { getImageUrl } from "./ImageList";
 // import zoom in from tabler
 import { IconZoomIn } from "@tabler/icons";
 import { SdImageBadgeBar } from "./SdImageBadgeBar";
+import { SdVariantHandler } from "./SdCardOrTableCell";
 
 // nextjs image
 type SdImageCompProps = {
@@ -21,6 +22,7 @@ type SdImageCompProps = {
   isMainImage?: boolean;
 
   onSetMainImage?(): void;
+  onCreateVariant?: SdVariantHandler;
 };
 
 export function SdImageComp(props: SdImageCompProps) {
@@ -52,6 +54,12 @@ export function SdImageComp(props: SdImageCompProps) {
               onSetMainImage={onSetMainImage}
               isMainImage={isMainImage}
             />
+            <Button onClick={() => props.onCreateVariant?.(image, "SD 1.5")}>
+              sd variant
+            </Button>
+            <Button onClick={() => props.onCreateVariant?.(image, "DALL-E")}>
+              DALL-E variant
+            </Button>
           </div>
         )}